temporary
refactor: :recycle: update deprecated options in both invocations of the command within the script

feat: :sparkles: add default preset for RuntimeGenesisConfig

Revert "feat: :sparkles: add default preset for RuntimeGenesisConfig"

This reverts commit 52436cb9e0a2a428022123d15bd0bb3f103ea009.

feat: :sparkles: add genesis builder dummy development preset

refactor: :recycle: implement GenesisBuilder within each runtime

refactor: :recycle: move moonbase runtime test config to runtime

refactor: :art: improve structure

refactor: :recycle: move moonbeam runtime test config to runtime

refactor: :recycle: move moonriver runtime test config to runtime

fix: :fire: remove unused dep

chore: :page_facing_up: add copyright

refactor: :fire: complete move of testnet_genesis functions to runtime

refactor: :coffin: remove dead functions *_inflation_config

refactor: :recycle: add get_preset and preset_names functions

build: :sparkles:

refactor: :recycle: correctly use with_genesis_config_preset_name in client code

refactor: :recycle: use with_genesis_config_preset_name in test_spec

refactor: :fire: remove serde_json feature

fix json serialization of GenesisConfig in xcm-transactor

fix serde_json dependendency in no_std environemnt

simplify code
