+++ client/evm-tracing/Cargo.toml
@@ -11,7 +11,7 @@ version = "0.1.0"
 ethereum-types = { workspace = true, features = [ "std" ] }
 hex = { workspace = true, features = [ "serde" ] }
 serde = { workspace = true, features = [ "derive", "std" ] }
-serde_json = { workspace = true }
+serde_json = { workspace = true, default-features = true }
 
 # Moonbeam
 evm-tracing-events = { workspace = true, features = [ "std" ] }
