+++ client/rpc-core/debug/Cargo.toml
@@ -15,7 +15,7 @@ jsonrpsee = { workspace = true, features = [ "macros", "server" ] }
 moonbeam-client-evm-tracing = { workspace = true }
 moonbeam-rpc-core-types = { workspace = true }
 serde = { workspace = true, features = [ "derive" ] }
-serde_json = { workspace = true }
+serde_json = { workspace = true, default-features = true }
 
 sp-core = { workspace = true, features = [ "std" ] }
 fc-rpc-core = { workspace = true }
