+++ client/rpc-core/trace/Cargo.toml
@@ -14,4 +14,4 @@ jsonrpsee = { workspace = true, features = [ "macros", "server" ] }
 moonbeam-client-evm-tracing = { workspace = true }
 moonbeam-rpc-core-types = { workspace = true }
 serde = { workspace = true, features = [ "derive" ] }
-serde_json = { workspace = true }
+serde_json = { workspace = true, default-features = true }
