+++ client/rpc-core/types/Cargo.toml
@@ -10,4 +10,4 @@ version = "0.1.0"
 [dependencies]
 ethereum-types = { workspace = true, features = [ "std" ] }
 serde = { workspace = true, features = [ "derive" ] }
-serde_json = { workspace = true }
+serde_json = { workspace = true, default-features = true }
