+++ runtime/moonriver/Cargo.toml
@@ -14,6 +14,7 @@ log = { workspace = true }
 num_enum = { workspace = true }
 rlp = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true, default-features = false, features = ["alloc"] }
 sha3 = { workspace = true, optional = true }
 smallvec = { workspace = true }
 strum = { workspace = true }
@@ -294,6 +295,7 @@ std = [
 	"parity-scale-codec/std",
 	"precompile-utils/std",
 	"scale-info/std",
+	"serde_json/std",
 	"session-keys-primitives/std",
 	"sp-api/std",
 	"sp-block-builder/std",
